Move script type detection to its own module
